requirements.txt
python-telegram-bot==20.7
mistralai==0.4.2
python-dotenv





